abstract-query/requirements.txt
git+https://github.com/morganwahl/zoneinfo
sqlparse>=0.4.3
typing_extensions>=4.3.0
mypy>=1.3.0
automathon>=0.0.13